Moved clean_sh, new factorials
